src/website/api/requirements.txt

azure-functions
azure-search-documents
azure-core